Fix copyrights statement
This started as a rewrite of mongodb-adapter
( https://github.com/casbin/mongodb-adapter ). As it turned out
this was not trivial rewrite and currently there is no share of
between these adapters. The only common things is general idea
and naming of things eg. default fields used by database.
Thus change (fix) of copyrights, license stays the same: APL2.0.